tidy up webpack config, basic port to typescript and add util for inlining config
