+++ Gruntfile.js
@@ -1,15 +1,6 @@
 module.exports = function (grunt) {
-	require('grunt-dojo2').initConfig(grunt, {
-		copy: {
-			staticDistFiles: { src: 'src/webpack.config.js', dest: 'dist/umd/webpack.config.js' },
-			staticDevFiles: { src: 'src/webpack.config.js', dest: '_build/src/webpack.config.js' }
-		}
-	});
-
+	require('grunt-dojo2').initConfig(grunt, {});
 	grunt.registerTask('ci', [
 		'intern:node'
 	]);
-
-	grunt.registerTask('dist', grunt.config.get('distTasks').concat(['copy:staticDistFiles']));
-	grunt.registerTask('dev', grunt.config.get('devTasks').concat(['copy:staticDevFiles']));
 };
